Laptop progress

+++ Shizzle_Logic/PostService.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Security;
 using System.Text;
 using System.Threading.Tasks;
+using System.Data;
 
 namespace Shizzle.Logic
 {
@@ -15,25 +16,72 @@ namespace Shizzle.Logic
         public uint authorityId { get; set; }
         private IPostDataService dataService;
         private IUserDataService userDataService;
-
-        public PostService(IPostDataService dataService)
+        private IGroupDataService groupDataService;
+        public PostService(IPostDataService dataService, IUserDataService userDataService, IGroupDataService groupDataService)
         {
             this.dataService = dataService;
+            this.userDataService = userDataService;
+            this.groupDataService = groupDataService;
         }
         public Structures.IPost CreatePost(string title, string content)
         {
+            title = title.Trim();
+            content = content.Trim();
+
+            if (string.IsNullOrEmpty(title) || string.IsNullOrEmpty(content))
+                throw new ArgumentException();
+
             return dataService.CreatePost(title, content, authorityId);
         }
 
         public Structures.IPost CreatePost(string title, string content, uint groupId)
         {
-            throw new NotImplementedException();
+            title = title.Trim();
+            content = content.Trim();
+
+            if (string.IsNullOrEmpty(title) || string.IsNullOrEmpty(content))
+                throw new ArgumentException();
+
+            bool isMember = false;
+            foreach(IGroup group in groupDataService.GetGroupsByUserParticipation(groupId))
+            {
+                if(group.id == groupId)
+                {
+                    isMember = true;
+                    break;
+                }
+            }
+
+            if (!isMember)
+                throw new SecurityException();
+
+            return dataService.CreatePost(title, content, authorityId, groupId);
         }
 
         public void DeletePost(uint id)
         {
-            //TODO: Check if you're the owner, or an admin of the group post
-            dataService.DeletePost(id);
+            IPost post = dataService.GetPost(id);
+
+            if(post is IGroupPost)
+            {
+                IGroup group = groupDataService.GetGroup(post.id);
+
+                if (group.adminIds.Contains(authorityId) || group.ownerId == authorityId)
+                {
+                    dataService.DeletePost(id);
+                    return;
+                }
+            } else
+            {
+                if (post.authorId == authorityId)
+                {
+                    dataService.DeletePost(id);
+                    return;
+                }
+            }
+
+            throw new SecurityException();
+
         }
 
         public void EditContent(uint id, string content)
